+++ CHANGELOG.md
@@ -4,6 +4,11 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [6.0.0-M14] - 2019-04-16
+### Changed
+- Renamed JdbcRepositoryHelper to JdbcResultSetStreamer
+- Extracted preparedStatementWrapperOf() in JdbcRepositoryHelper to a new class 'PreparedStatementWrapperFactory'
+
 ## [6.0.0-M13] - 2019-04-16
 ### Changed
 - Upgrade to latest framework-api version 4.0.0-M8